fix benchmarks and cookbooks problems
fix various warnings and errors in cookbooks and benchmarks

+++ benchmarks/crameri_et_al/case_1/crameri_benchmark_1.cc
@@ -80,7 +80,7 @@ namespace aspect
 
       private:
 
-        unsigned int order;  //Order of the perturbation
+        double order;  //Order of the perturbation
         double amplitude;  //amplitude of the perturbation
 
     };
